services/send-to-intercom/requirements.txt
markdown>=3.3.4
python-frontmatter>=0.5.0
python-dotenv>=0.19.0
requests>=2.26.0